add: reverse search in ban api

+++ src/main/java/net/botlify/ban/api/param/SearchAddressParam.java
@@ -35,10 +35,10 @@ public class SearchAddressParam {
   public static class Position {
 
     @Getter
-    private final long latitude;
+    private final double latitude;
 
     @Getter
-    private final long longitude;
+    private final double longitude;
 
   }
